feat: add surface Laplacian, resolution metrics, label time course, SPoC (batch 10)
New classes:
- SurfaceLaplacian: Current Source Density via spherical spline (Perrin 1989)
- InvResolutionMatrix: PSF/CTF analysis, spatial spread, peak localisation error
- InvLabelTimeCourse: extract label time courses (mean/mean_flip/pca_flip/max/auto)
- MlSpoc: Source Power Comodulation spatial filter (generalised eigenvalue)

Also fixes:
- Format string warnings in eeg_reference.cpp and bad_channels_maxwell.cpp
- Missing override specifiers in covariance plugin header

Tests: 51 new (12+14+13+12), running total 295

+++ src/libraries/dsp/bad_channels_maxwell.cpp
@@ -156,8 +156,8 @@ BadChannelsMaxwellResult UTILSLIB::findBadChannelsMaxwell(
         }
     }
 
-    qInfo("[findBadChannelsMaxwell] Detected %d bad channel(s) out of %d MEG channels.",
-          result.badChannels.size(), nMeg);
+qInfo("[findBadChannelsMaxwell] Detected %lld bad channel(s) out of %d MEG channels.",
+           static_cast<long long>(result.badChannels.size()), nMeg);
 
     return result;
 }
